Use COPR repo for EL7 install
The upstream BIRD RPMs do not support installation on EL7 systems,
they are built for Fedora only and the dependencies are not present
on EL7 repos or EPEL.

To get EL7 installation working, I created a COPR repo to create
automatic RPM builds for EL6 and EL7 systems based on the upstream
SRPM files.

The COPR repo is available here:
https://copr.fedorainfracloud.org/coprs/logan2211/bird/ and will be
used as the default BIRD repo for EL7 installations.

EPEL installation is still available however the version in EPEL is
so outdated it is probably insufficient for many modern uses of BIRD.